+++ packages/common/src/types.ts
@@ -1,0 +1,9 @@
+import { Blueprint, Predicates } from 'optimal';
+
+export interface Optionable<T extends object = {}> {
+  options: Required<T>;
+
+  blueprint(predicates: Predicates): Blueprint<T>;
+
+  configure(options: Partial<T>): this;
+}