+++ packages/core/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Added a new package, [@boost/common](https://www.npmjs.com/package/@boost/common), to house common
   utilities and helpers for building Boost or general purpose applications.
-  - Added an `Optionable` abstract class.
+  - Added a `Contract` abstract class.
 - Added a new package, [@boost/event](https://www.npmjs.com/package/@boost/event), to provide a
   type-safe static event system. The old event emitter is deprecated, so please migrate to the new
   system!